Fix conio.h problem on Solaris Add debug code to lock_jcr_chain Implement call by reference for Mmsg() and pm_strxxx()
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1541 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/wx-console/wxbmainframe.cpp
@@ -569,7 +569,7 @@ void wxbMainFrame::Print(wxString str, int status)
       wxbDataTokenizer* dt = wxbUtils::WaitForEnd(".help", true);
       int i, j;
       wxString str;
-      for (i = 0; i < dt->GetCount(); i++) {
+      for (i = 0; i < (int)dt->GetCount(); i++) {
          str = (*dt)[i];
          str.RemoveLast();
          if ((j = str.Find(' ')) > -1) {
